Change transparent background back to white.
Originally I wanted to avoid using a pure white background to
avoid ugliness when displayed on top of an off-white or other light
color background, but ultimately these diagrams end up in Teams and
Confluence, which use pure white, so that was a concern that didn't
really matter in practice. On the other hand, a transparent background
is problematic when embedded in a dark mode context, as the choice of
dark colors for text results in unreadable diagrams. Supporting dark
mode can't be done in an automatic way, but should be doable as a
separate theme.